backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "ezdxf>=1.4.3",
    "fastapi>=0.129.0",
    "openpyxl>=3.1.5",
    "ortools>=9.15.6755",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "razorpay>=2.0.0",
    "reportlab>=4.4.10",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
